Python Project 1
 Create a  Python Script with the bellow executables:
Allow the user to input how many EC2 instances they want names for and output the same amount of unique names.
Allow the user to input the name of their department that is used in the unique name.
Generate random characters and numbers that will be included in the unique name.
Push your code to GitHub and include the link in your LinkedIn write up.