Update pythonproject1.py
This program request the user to input the number of EC2 instances desired and the department name. if the user inputs the wrong department name they receive an error message and are asked the department name again. If the input is correct, the program prints the number of EC2 instance with a random assigned integer attached to the instance name.